Move EARExceptions into ear-core
