+++ ear-compiler/src/main/java/eflang/ear/EARCompiler.java
@@ -3,6 +3,7 @@ package eflang.ear;
 import eflang.ear.composer.Composer;
 import eflang.ear.composer.OnlyRunsComposer;
 import eflang.ear.core.Command;
+import eflang.ear.core.EARException;
 import eflang.ear.core.Scales;
 import eflang.ear.core.StatefulInstructionCompiler;
