+++ ear-core/src/main/java/eflang/ear/core/EARException.java
@@ -1,4 +1,4 @@
-package eflang.ear;
+package eflang.ear.core;
 
 public class EARException extends Exception {
     EARException() {
