+++ ear-core/src/main/java/eflang/ear/core/EARInvalidOpcodeException.java
@@ -1,4 +1,4 @@
-package eflang.ear;
+package eflang.ear.core;
 
 class EARInvalidOpcodeException extends EARException {
     EARInvalidOpcodeException(String message) {
