+++ ear-core/src/main/java/eflang/ear/core/EARInvalidSignatureException.java
@@ -1,4 +1,4 @@
-package eflang.ear;
+package eflang.ear.core;
 
 class EARInvalidSignatureException extends EARException {
     EARInvalidSignatureException(String message) {
