+++ vibe/src/main/java/eflang/vibe/VibeController.java
@@ -3,7 +3,7 @@ package eflang.vibe;
 import eflang.core.Parser;
 import eflang.core.StringMusicSource;
 import eflang.ear.EARCompiler;
-import eflang.ear.EARException;
+import eflang.ear.core.EARException;
 import eflang.ear.core.Scale;
 import eflang.ear.core.Scales;
 import eflang.ear.composer.GeometricComposer;
